251127 벡엔드 - 지출금액 amount type Long으로 변경 - 지출 통계 API service, controller 작성 - queryDsl과 PostgreSQL 문법 충돌로 인한 에러 해결

+++ FinTrack_BE/src/main/java/doHoaSen/FinTrack/expense/entity/Expense.java
@@ -22,7 +22,7 @@ public class Expense {
     @ManyToOne(fetch = FetchType.LAZY)
     private User user;
 
-    private Integer amount;
+    private Long amount;
     private String category;
     private String memo;
     private LocalDateTime dateTime;
